refactor: 관심사에 따라 controller 분리
