+++ src/domain/LottoCalculator.js
@@ -26,6 +26,7 @@ class LottoCalculator {
   #calculateRateOfRevenue(revenue, lottosCount) {
     return ((revenue / (lottosCount * NUMBER.LOTTO_PRICE)) * 100).toFixed(1);
   }
+
 }
 
 export default LottoCalculator;
